Added state and navigation
There are now three different screens that the user can navigate between by clicking buttons.
Each screen is represented by a game state. Each screen has a draw and update function.
Switch statements are used to handle drawing and updating for each state.
Added mouseover functionality for a button so it highlights when the cursor is over it.