Import the Ably Go SDK and start using the realtime client.
Three buttons on the realtime screen have been created.
These buttons create a client, get a channel and publish a message.
At the moment the buttons turn green if operations were successful.
A white bar has also been added to the top of this screen.
Planning to use this bar to to display feedback to the user.