Fixed a bug where holding mouse button created endless clients
The create button for the realtime client was checking if a client
existed before attempting to create a client. This resulted in many
clients being created while the mouse button was held down.
This bug was identified after the platform sent me an email saying
that peak connection hard limit had been exceeded. I have started
to add some logging to the project. I've also started to keep a list
of potential SDK issues found in the readme file.