Display channel information
It is now possible to set a channel on a connection.
The channel is displayed in a yellow box inside the green client box.
Have created a new text package for drawing text to the screen.
This is used to display the channel information as using debug prints
to display information felt unsustainable.
Curently a default channel name is used.