Added subscribe all functionality to a channel.
Once a client has been created and a channel has been set, a button
labelled subscribe all appears. Clicking on this button will subscribe
the channel to all messages. The unsubscribe function returned by the
SDK is also saved for future consumption. As the status of the channel
is displayed on the screen, this can also been seen to change from
initialised to attached. Have decided to reuse the subscribe button for
unsubscribing so its colour and text also change when it is clicked.