Added functionality to get presence for a channel.
The call to get presence is an async go routine. At the moment it
is possible to get presence on a channel that has been initialised.
When doing this, sometimes the call to get presence succeeds and
sometimes it fails. It feels like there could be a race condition bug
in this area. Have updated the readme with observations to investigate
further.