Added functionality to remove presence (leave).
I have also re-ordered the UI elements for announce, get and leave.
This is so they follow a CRUD (create, read, update delete) order.
Announce creates so it is first, then get which reads is next with
leave which deletes appearing last. Its possible to set presence data
in the ably platform and update presence data, but I have left presence
data out of scope for the context of this demo.