Added functionality to publish to a channel, messages now displayed
Subscribing to a channel now connects to an event info text box on the
screen. When an event occurs on the channel, the event name and data are
drawn to the event info element on the screen. Event info gets torn down
when a channel is unsubscribed from or when a client is closed.