Implemented text input box functionality for setting channel name.
Have created a new package called textbox which creates textbox
elements that are capable of capturing user input from the keyboard.
An area labelled inputs is now drawn on the ui with a textbox that
captures a channel name. The name in the input box is used to create
the channel. This means its now possible to have more than one channel
created.