Change hardcoded location of close client button
The close client buttons are now drawn relative to the create button.
This makes the only location which is hard coded the location of the
create button. This should open the door to initialising all elements
with a single function that knows the connectionID and location of that
connections create button.