Add more information to the title screen including SDK version
On initialisation a bash command is run to extract the SDK version
of ably-go from the go.mod file located at the project root directory.
The go list command is used to get this information and it is stored in
the config global. This raised the question as to whether ably client
libraries should expose this information to developers by default.
A git hub issue has been created for this, see
https://github.com/ably/docs/issues/1291